Rename gpytorch_gaussian_process module to gaussian_process
The gpytorch prefix exposed an implementation detail. The module now has a
cleaner name that focuses on what it does (Gaussian Process) rather than
how it does it (using GPyTorch library).

Also removed unused scale_x and normalize_y parameters from plot_test_predictions.

Updated all imports across tests, scripts, and the bayesian_optimization module.
