+++ scripts/gaussian_process_regression/gp_fromdata.py
@@ -20,7 +20,7 @@ from sklearn.metrics import mean_absolute_error, mean_squared_error
 
 from surmod import data_processing
 
-from surmod.gpytorch_gaussian_process import GPSurrogate
+from surmod.gaussian_process import GPSurrogate
 
 
 def parse_arguments():
