+++ scripts/gaussian_process_regression/gp_sandbox.py
@@ -39,7 +39,7 @@ from sklearn.metrics import mean_absolute_error, mean_squared_error
 
 from surmod.test_functions import simulate_data
 
-from surmod.gpytorch_gaussian_process import GPSurrogate
+from surmod.gaussian_process import GPSurrogate
 
 
 def parse_arguments():
