+++ scripts/sensitivity_analysis/sa_fromdata.py
@@ -41,7 +41,7 @@ from sklearn.metrics import mean_absolute_error, mean_squared_error as mse
 
 from surmod import sensitivity_analysis as sa, data_processing
 
-from surmod.gpytorch_gaussian_process import GPSurrogate
+from surmod.gaussian_process import GPSurrogate
 
 
 def parse_arguments():
